added a file for export_utils.c && moved func to it
